# record-label-api/openapi.yaml
openapi: 3.0.0
info:
  title: Record Label API
  version: 1.0.0
  description: "A simple record label API with Basic Authentication and pagination."
servers:
  - url: http://localhost:8000
paths:
  /artists:
    get:
      summary: List artists with pagination
      security:
        - basicAuth: []
      parameters:
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        '200':
          description: A paginated list of artists
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Artist'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
    post:
      summary: Create a new artist
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Artist'
      responses:
        '201':
          description: Artist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
  /artists/{artistname}:
    get:
      summary: Get a specific artist by username
      security:
        - basicAuth: []
      parameters:
        - name: artistname
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Artist found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    Artist:
      type: object
      required:
        - name
        - genre
        - albums
        - username
      properties:
        name:
          type: string
        genre:
          type: string
        albums:
          type: integer
        username:
          type: string